Separate authorization and token endpoints
