+++ src/AuthorizationMiddlewareFactory.php
@@ -13,26 +13,16 @@ namespace Zend\Expressive\Authentication\OAuth2;
 use League\OAuth2\Server\AuthorizationServer;
 use Psr\Container\ContainerInterface;
 use Psr\Http\Message\ResponseInterface;
-
 use function sprintf;
 
-class OAuth2MiddlewareFactory
+class AuthorizationMiddlewareFactory
 {
-    public function __invoke(ContainerInterface $container) : OAuth2Middleware
+    use AuthFlowFactoryTrait;
+
+    public function __invoke(ContainerInterface $container) : AuthorizationMiddleware
     {
-        $authServer = $container->has(AuthorizationServer::class)
-            ? $container->get(AuthorizationServer::class)
-            : null;
-
-        if (null === $authServer) {
-            throw new Exception\InvalidConfigException(sprintf(
-                "The %s service is missing",
-                AuthorizationServer::class
-            ));
-        }
-
-        return new OAuth2Middleware(
-            $authServer,
+        return new AuthorizationMiddleware(
+            $this->getAuthorizationServer($container),
             $container->get(ResponseInterface::class)
         );
     }
